requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.1
Django==1.11.13
djangorestframework==3.8.2
kombu==4.2.0
psycopg2==2.7.4
pytz==2018.4
redis==2.10.6
vine==1.1.4